Enhance renderer selection for improved browser compatibility
WebGL2 support across browsers if far from uniform and varies greatly between browsers and their many versions. Ttyd attempts to use WebGL2 as the default renderer unless the user statically specifies 'rendererType' via the client options, There is existing logic to fail back to canvas in the event that WebGL cannot be loaded by the client browser. However, many older browsers lacking WebGL2 do not seem to produce an error which is caught by 'enableWebglRenderer' and ttyd will end up loading the page in a broken way without applying any selected client options due to 'applyPreferences' not finishing its execution.

To handle such cases in a more elegant manner, I added a function called 'setRendererTypeSafely' which replaces the call to 'setRendererTypeSafely' in 'applyPreferences'. For most browsers lacking WebGL2 support, we can simply test the client for WebGL2 support by attempting to get a WebGL2 rendering context from a canvas element. If we fail to get the context we log lack of WebGL2 support to the console then gracefully fall back to canvas as the renderer. This is what 'setRendererTypeSafely' does for most browsers.

When dealing with versions of Safari older than 16.x, the simple rendering context test to determine browser support of WebGL2 does not work as Apple has made any attempt at requesting WebGL elements result in a fatal error which instantly breaks the page load and prevents 'applyPreferences' from continuing. To gracefully handle this edge case, 'setRendererTypeSafely' will check for Safari before performing any WebGL2 context tests and if Safari is found to be older than 16, it will fall back to canvas as the renderer.

There are many scenarios in which devices have not been able to update their browser versions so this should make ttyd much more usable for them. With the addition of this function, browser support is greatly enhanced as many older browser versions can now connect to ttyd without breaking the selected client options.