html: upgrade to xterm 5.4.0 (#1303)
* html: Upgrade xterm.js

Upgrade xterm.js to the latest stable release.

As part of this upgrade, the xterm packages have been scoped as the
unscoped packages were not updated.

The addons have also been upgraded to match the compatible versions.

As part of this upgrade, a compatibility issue was identified with the
Gulp task to inline source. By disabling compression, this allowed the
task to complete successfully. The size of the source file `html.h` did
not change significantly which indicates compression was not being
applied previously.

Signed-off-by: Michael Lorant <michael.lorant@nine.com.au>

* html: rebuild header file

---------

Signed-off-by: Michael Lorant <michael.lorant@nine.com.au>
Co-authored-by: Shuanglei Tao <tsl0922@gmail.com>